+++ packages/core/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact core module, newest changes on the top.
 
+## [1.14.0] - 2017-02-23
+
+### Deprecated
+
+- `core/factory`, to be removed in 2.0.0
+
 ## [1.13.3] - 2017-01-16
 
 ### Fixed